refactor SSO re-auth
